+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### [Unreleased][unreleased]
 
+#### Changed
+- split inner libraries and their corresponding tests into their own files
+
 ### 0.1.0 - 2015-09-28
 #### Added
 - package files