+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### [Unreleased][unreleased]
 
+### 0.2.0 - 2015-09-29
 #### Removed
 - QUICKMOCK_STUBS global var in favor of a more generic QUICKMOCK_DOUBLES
 - QuickMock.trackStub in favor of a more generic QuickMock.trackDouble